deleted assignments App and migrated all functionalities using assignmentsapp to MCQApp
