+++ BTalaqa/BTalaqa/settings.py
@@ -39,8 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'MCQAssignmentsApp',
-    'AssignmentsApp'
+    'MCQAssignmentsApp'
 ]
 
 MIDDLEWARE = [
