+++ BTalaqa/BTalaqa/templates/common/base.html
@@ -47,7 +47,7 @@
                     <a class="nav-link " href={% url 'MCQAssignmentsApp:create_test' %}> <i class="fa fa-file"></i> Create Test<span class="sr-only">(current)</span></a>
                   </li>
                   <li class="nav-item active">
-                    <a class="nav-link" href={% url 'AssignmentsApp:assign_users_tests' %}> <i class="fas fa-hands-helping"></i> Assign Tests<span class="sr-only">(current)</span></a>
+                    <a class="nav-link" href={% url 'MCQAssignmentsApp:assign_users_tests' %}> <i class="fas fa-hands-helping"></i> Assign Tests<span class="sr-only">(current)</span></a>
                   </li>
                   <li class="nav-item active">
                     <a class="nav-link" href={% url 'MCQAssignmentsApp:dashboard' %}><i class="fas fa-tachometer-alt"></i> Dashboard<span class="sr-only"></span></a>
